0.0.042

+++ luos/thread/english/available/internal/variable.inl
@@ -15,6 +15,6 @@ std::atomic<size_t> draw_completed{0};
 std::atomic<size_t> update_outstanding{0};
 std::atomic<size_t> draw_outstanding{0};
 
-uint32_t current_gen{0};
-uint32_t update_done_gen{0};
-uint32_t draw_done_gen{0};
+uint64_t current_gen{0};
+uint64_t update_done_gen{0};
+uint64_t draw_done_gen{0};
