+++ luos/window/english/parameters/window_vulkan/english/parameters/command_buffer/english/available/variable.inl
@@ -1,4 +1,4 @@
 uint32_t viewport_width{0};
 uint32_t viewport_height{0};
-uint32_t cut_offset_width{0};
-uint32_t cut_offset_height{0};
+uint16_t cut_offset_width{0};
+uint16_t cut_offset_height{0};
